remove indexedmap from contract + some tweaks
