+++ contracts/ecash/src/lib.rs
@@ -3,7 +3,6 @@
 
 pub mod contract;
 pub mod errors;
+mod helpers;
 #[cfg(test)]
 pub mod multitest;
-mod state;
-mod storage;
